fixs: updated the reportparserservice.java and removed redundant data

+++ Backend/tverseIQ/src/main/java/com/tverseIQ/backend/dto/ParsedRowDto.java
@@ -8,7 +8,7 @@ public record ParsedRowDto (
         String keyword,
         String matchType,
         BigDecimal spend,
-        BigDecimal orders,
+        Integer orders,
         BigDecimal sales
 
 ){}
